lock numpy==1.22.1 and make toScreen()->float

+++ friture/plotting/coordinateTransform.py
@@ -20,7 +20,7 @@
 from PyQt5 import QtCore
 from PyQt5.QtCore import pyqtSlot
 import numpy as np
-import friture.plotting.frequency_scales as fscales 
+import friture.plotting.frequency_scales as fscales
 
 # transforms between screen coordinates and plot coordinates
 
@@ -58,31 +58,31 @@ class CoordinateTransform(QtCore.QObject):
     def setScale(self, scale):
         self.scale = scale
 
-    @pyqtSlot(float, result=int)
+    @pyqtSlot(float, result=float)
     def toScreen(self, x):
         if self.scale is fscales.Logarithmic:
             if self.coord_clipped_min == self.coord_clipped_max:
-                return int(self.startBorder + 0. * x)  # keep x type (this can produce a RunTimeWarning if x contains inf)
+                return self.startBorder + 0. * x  # keep x type (this can produce a RunTimeWarning if x contains inf)
 
             x = (x < 1e-20) * 1e-20 + (x >= 1e-20) * x
-            return int((np.log10(x / self.coord_clipped_min))
-                       * (self.length - self.startBorder - self.endBorder)
-                       / self.coord_ratio_log
-                       + self.startBorder)
+            return (np.log10(x / self.coord_clipped_min)
+                    * (self.length - self.startBorder - self.endBorder)
+                    / self.coord_ratio_log
+                    + self.startBorder)
         else:
             if self.coord_max == self.coord_min:
-                return int(self.startBorder + 0. * x)  # keep x type (this can produce a RunTimeWarning if x contains inf)
-            
+                return self.startBorder + 0. * x  # keep x type (this can produce a RunTimeWarning if x contains inf)
+
             trans_x = self.scale.transform(x)
             trans_min = self.scale.transform(self.coord_min)
             trans_max = self.scale.transform(self.coord_max)
 
-            return int((trans_x - trans_min)
-                       * (self.length - self.startBorder - self.endBorder)
-                       / (trans_max - trans_min)
-                       + self.startBorder)
+            return ((trans_x - trans_min)
+                    * (self.length - self.startBorder - self.endBorder)
+                    / (trans_max - trans_min)
+                    + self.startBorder)
 
-    @pyqtSlot(int, result=float)
+    @pyqtSlot(float, result=float)
     def toPlot(self, x):
         if self.length == self.startBorder + self.endBorder:
             return self.coord_min + 0. * x  # keep x type (this can produce a RunTimeWarning if x contains inf)
